agent_knowledge/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
chromadb>=0.5.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
requests>=2.31.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
openai>=1.30.0
google-generativeai>=0.8.0
